requirements.txt
discord.py-self==2.1
google-generativeai==0.3.2
anthropic==0.8.1
deepseek==1.0.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
protobuf>=3.20.0,<5.0.0
supabase==2.0.3
